requirements.txt
# Core dependencies
numpy>=1.24.3
pandas>=2.0.2
scikit-learn>=1.2.2
xgboost>=1.7.5
joblib>=1.2.0
Flask>=2.3.2
Werkzeug>=2.3.6
python-dotenv>=1.0.0
gunicorn>=20.1.0

# Visualization
matplotlib>=3.7.1
seaborn>=0.12.2
plotly>=5.14.1

# Development and testing
pytest>=7.3.1
black>=23.3.0
flake8>=6.0.0
pytest-cov>=4.0.0

# Jupyter notebook support
jupyter>=1.0.0
ipywidgets>=7.6.5

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0
